// src/apprunner/models/apprunner.d.ts
    export interface AuthenticationConfiguration {
        ConnectionArn?: FusionResourceArn
        AccessRoleArn?: RoleArn
    }
// ...
    export interface CodeRepository {
        RepositoryUrl: String
        SourceCodeVersion: SourceCodeVersion
        CodeConfiguration?: CodeConfiguration
    }
// ...
    export type Cpu = string
// ...
    export interface CreateServiceRequest {
        ServiceName: ServiceName
        SourceConfiguration: SourceConfiguration
        InstanceConfiguration?: InstanceConfiguration
        Tags?: TagList
        EncryptionConfiguration?: EncryptionConfiguration
        HealthCheckConfiguration?: HealthCheckConfiguration
        AutoScalingConfigurationArn?: FusionResourceArn
    }
// ...
    export interface EncryptionConfiguration {
        KmsKey: KmsKeyArn
    }
    export type FusionResourceArn = string
    export interface HealthCheckConfiguration {
        Enabled?: NullableBoolean
        Path?: String
        Interval?: HealthCheckInterval
        Timeout?: HealthCheckTimeout
        HealthyThreshold?: HealthCheckHealthyThreshold
        UnhealthyThreshold?: HealthCheckUnhealthyThreshold
    }
    export type HealthCheckHealthyThreshold = number
    export type HealthCheckInterval = number
    export type HealthCheckTimeout = number
    export type HealthCheckUnhealthyThreshold = number
    export interface ImageConfiguration {
        RuntimeEnvironmentVariables?: RuntimeEnvironmentVariables
        StartCommand?: String
        Port?: String
    }
    export type ImageIdentifier = string
    export interface ImageRepository {
        ImageIdentifier: ImageIdentifier
        ImageConfiguration?: ImageConfiguration
        ImageRepositoryType: ImageRepositoryType
    }
    export type ImageRepositoryType = 'ECR' | 'ECR-Public' | string
    export type InitiatorType = 'USER' | 'SERVICE' | string
    export interface InstanceConfiguration {
        Cpu?: Cpu
        Memory?: Memory
        InstanceRoleArn?: RoleArn
    }
    export type Integer = number
    export type KmsKeyArn = string
// ...
    export type Memory = string
    export type NextToken = string
    export type NullableBoolean = boolean
// ...
    export type RoleArn = string
// ...
    export type ServiceName = string
// ...
    export interface SourceConfiguration {
        CodeRepository?: CodeRepository
        ImageRepository?: ImageRepository
        AutoDeploymentsEnabled?: NullableBoolean
        AuthenticationConfiguration?: AuthenticationConfiguration
    }
    export interface StartDeploymentRequest {
        ServiceArn: FusionResourceArn
    }
    export interface StartDeploymentResponse {
        StartDeploymentResult: StartDeploymentResult
    }
    export interface StartDeploymentResult {
        OperationId: UUID
    }
    export type String = string
    export interface Tag {
        Key?: TagKey
        Value?: TagValue
    }
    export type TagKey = string
    export type TagKeyList = TagKey[]
    export type TagList = Tag[]