// types/git.ts
export const enum RefType {
    Head,
    RemoteHead,
    Tag,
}

export interface Ref {
    readonly type: RefType
    readonly name?: string
    readonly commit?: string
    readonly remote?: string
}

export interface UpstreamRef {
    readonly remote: string
    readonly name: string
}

export interface Branch extends Ref {
    readonly upstream?: UpstreamRef
    readonly ahead?: number
    readonly behind?: number
}

export interface Commit {
    readonly hash: string
    readonly message: string
    readonly parents: string[]
    readonly authorDate?: Date
    readonly authorName?: string
    readonly authorEmail?: string
    readonly commitDate?: Date
}

export interface Submodule {
    readonly name: string
    readonly path: string
    readonly url: string
}

export interface Remote {
    readonly name: string
    readonly fetchUrl?: string
    readonly pushUrl?: string
    readonly isReadOnly: boolean
}